requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
executing==1.2.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jsonpickle==3.0.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
networkx==3.1
numpy==1.25.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
prettytable==3.8.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pyvis==0.3.2
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6
